requirements.txt
django==3.2
mutagen==1.45.1h
django-storages==1.12.3
boto3==1.21.4
python-magic==0.4.25
python-environ==0.4.54
djangorestframework==3.13.1
django-url-filter==0.3.15
ffmpeg-python==0.2.0
celery==5.2.3